refactor: move projection pushing into its own file
